Refactor setup wizard into modular structure with golden rule fixes
Split pages.py (~2359 lines) into smaller, focused modules:
- utils.py (170 lines): Shared utility functions including new helpers
  for muted labels and metadata stylesheets
- welcome.py (151 lines): WelcomePage, DatabaseInstructionsPage
- database_config.py (310 lines): DatabaseConfigPage
- database_setup.py (629 lines): DatabaseSetupWorker, DatabaseSetupPage
- api_keys.py (263 lines): APIKeysPage
- import_options.py (262 lines): ImportOptionsPage
- import_progress.py (422 lines): ImportWorker, ImportProgressPage
- document_browser.py (285 lines): DocumentBrowserPage
- complete.py (139 lines): CompletePage

pages.py now serves as a compatibility layer re-exporting all classes.

Golden rule compliance fixes:
- Replace inline stylesheets with StylesheetGenerator (rule #9)
- Replace hardcoded pixel values with scale-based sizes (rule #10)
- Add FRAME_METADATA_BG, SPLITTER_LIST_RATIO, SPLITTER_PREVIEW_RATIO
  constants to constants.py (rule #2)
- Add create_muted_label_stylesheet() and create_metadata_label_stylesheet()
  utility functions

Note: database_setup.py (629 lines) slightly exceeds 500-line target
due to tightly coupled worker thread and page classes.
