+++ src/bmlibrarian/gui/qt/setup_wizard/__init__.py
@@ -6,6 +6,19 @@ This module provides a PySide6 QWizard-based setup interface for:
 2. Database schema initialization
 3. Optional API key configuration (Anthropic, OpenAI)
 4. Data source import (PubMed, medRxiv, MeSH)
+5. Document browser for verifying imports
+
+The pages are organized into individual modules for maintainability:
+- welcome.py: WelcomePage, DatabaseInstructionsPage
+- database_config.py: DatabaseConfigPage
+- database_setup.py: DatabaseSetupWorker, DatabaseSetupPage
+- api_keys.py: APIKeysPage
+- import_options.py: ImportOptionsPage
+- import_progress.py: ImportWorker, ImportProgressPage
+- document_browser.py: DocumentBrowserPage
+- complete.py: CompletePage
+- utils.py: Shared utility functions
+- constants.py: Configuration constants
 """
 
 from .wizard import SetupWizard
@@ -13,10 +26,13 @@ from .pages import (
     WelcomePage,
     DatabaseInstructionsPage,
     DatabaseConfigPage,
+    DatabaseSetupWorker,
     DatabaseSetupPage,
     APIKeysPage,
     ImportOptionsPage,
+    ImportWorker,
     ImportProgressPage,
+    DocumentBrowserPage,
     CompletePage,
 )
 
@@ -25,9 +41,12 @@ __all__ = [
     "WelcomePage",
     "DatabaseInstructionsPage",
     "DatabaseConfigPage",
+    "DatabaseSetupWorker",
     "DatabaseSetupPage",
     "APIKeysPage",
     "ImportOptionsPage",
+    "ImportWorker",
     "ImportProgressPage",
+    "DocumentBrowserPage",
     "CompletePage",
 ]
