+++ src/bmlibrarian/gui/qt/setup_wizard/constants.py
@@ -62,3 +62,9 @@ FRAME_WARNING_BG = "#FFEBEE"
 FRAME_WARNING_BORDER = "#EF5350"
 FRAME_SUCCESS_BG = "#E8F5E9"
 FRAME_SUCCESS_BORDER = "#66BB6A"
+FRAME_METADATA_BG = "#F5F5F5"
+FRAME_METADATA_BORDER = "#E0E0E0"
+
+# Splitter size ratios (as percentages)
+SPLITTER_LIST_RATIO = 40  # Document list takes 40% of width
+SPLITTER_PREVIEW_RATIO = 60  # Preview takes 60% of width
